Fixed_structures_grammatical_errors.md
Changed "Look no further as this page shows you the player! all of the current structures you can find within the Add-On" to "Look no further, as this page shows you, the player, all of the current structures you can find within the Add-On!" to make it grammatically correct while adding an exclamation mark for engagement.

Replaced unneeded coma with an em dash to "Any traveler that goes near it—will have to battle for peace" to make it grammatically correct while keeping the suspenseful pause. If desired, you can remove the em dash if you want to remove the pause in the sentence, but putting a coma there wouldn't be recommended by me.

Made adventurer camp more consistent in past tense. "This structure is home to one lost traveler. It was set up at random places in a world. It was his way to write a story." Is changed to "This structure was home to one lost traveler. He set them up in random places throughout the world. It was his way to write a story."

Fixed "Huh A way back to the overworld. Cool!" While also making it a question "Huh? A way back to the overworld? Cool"

Thought about adding a question mark to "huh, I wonder why they left.." but with three dots at the end it is fully optional, so I left it untouched.